services/visual_analysis/app/core/model.py
"""
DETR model implementation for fashion analysis
TODO: Implement actual DETR model
"""

import torch
import torch.nn as nn
from transformers import DetrImageProcessor, DetrForObjectDetection
from typing import Dict, List, Any, Tuple
import numpy as np
from PIL import Image

class DETRFashionModel:
    """DETR model for fashion image analysis"""
    
    def __init__(self, model_name: str = "facebook/detr-resnet-50"):
        """Initialize DETR model"""
        self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
        self.processor = DetrImageProcessor.from_pretrained(model_name)
        self.model = DetrForObjectDetection.from_pretrained(model_name)
        self.model.to(self.device)
        self.model.eval()
        
        # Fashion-specific class mappings (to be customized for Fashionpedia)
        self.fashion_classes = {
            0: "shirt",
            1: "pants", 
            2: "dress",
            3: "skirt",
            4: "jacket",
            5: "shoes",
            # Add more mappings based on Fashionpedia dataset
        }
    
    def preprocess_image(self, image: Image.Image) -> Dict[str, torch.Tensor]:
        """Preprocess image for DETR model"""
        inputs = self.processor(images=image, return_tensors="pt")
        return {k: v.to(self.device) for k, v in inputs.items()}
    
    def predict(self, image: Image.Image, confidence_threshold: float = 0.5) -> Dict[str, Any]:
        """
        Predict fashion items in image
        
        Args:
            image: PIL Image
            confidence_threshold: Minimum confidence for detections
            
        Returns:
            Dictionary with detected items and metadata
        """
        # Preprocess image
        inputs = self.preprocess_image(image)
        
        # Run inference
        with torch.no_grad():
            outputs = self.model(**inputs)
        
        # Post-process results
        target_sizes = torch.tensor([image.size[::-1]]).to(self.device)  # (height, width)
        results = self.processor.post_process_object_detection(
            outputs, 
            target_sizes=target_sizes, 
            threshold=confidence_threshold
        )[0]
        
        # Convert to fashion analysis format
        detected_items = []
        for score, label, box in zip(results["scores"], results["labels"], results["boxes"]):
            category = self.fashion_classes.get(label.item(), "unknown")
            
            # Normalize coordinates
            x1, y1, x2, y2 = box.cpu().numpy()
            width, height = image.size
            
            detected_items.append({
                "category": category,
                "confidence": score.item(),
                "bounding_box": {
                    "x": x1 / width,
                    "y": y1 / height,
                    "width": (x2 - x1) / width,
                    "height": (y2 - y1) / height
                },
                # TODO: Add color, style, pattern detection
                "color": "unknown",
                "style": "unknown", 
                "pattern": "unknown"
            })
        
        return {
            "items": detected_items,
            "metadata": {
                "model_name": "detr-fashion",
                "confidence_threshold": confidence_threshold,
                "device": str(self.device)
            }
        }
    
    def extract_features(self, image: Image.Image) -> np.ndarray:
        """Extract feature embeddings from image"""
        inputs = self.preprocess_image(image)
        
        with torch.no_grad():
            outputs = self.model(**inputs, output_hidden_states=True)
            # Extract features from last hidden state
            features = outputs.last_hidden_state.mean(dim=1)  # Global average pooling
            
        return features.cpu().numpy()


class FashionAttributeClassifier:
    """Classifier for fashion attributes (color, style, pattern)"""
    
    def __init__(self):
        """Initialize attribute classifiers"""
        # TODO: Implement actual classifiers
        self.color_classifier = None
        self.style_classifier = None
        self.pattern_classifier = None
    
    def predict_color(self, image_crop: Image.Image) -> str:
        """Predict color of clothing item"""
        # TODO: Implement color classification
        return "blue"
    
    def predict_style(self, image_crop: Image.Image) -> str:
        """Predict style of clothing item"""
        # TODO: Implement style classification
        return "casual"
    
    def predict_pattern(self, image_crop: Image.Image) -> str:
        """Predict pattern of clothing item"""
        # TODO: Implement pattern classification
        return "solid"


class SegmentationModel:
    """Model for generating segmentation masks"""
    
    def __init__(self):
        """Initialize segmentation model"""
        # TODO: Implement segmentation model (e.g., using DETR with masks)
        pass
    
    def generate_mask(self, image: Image.Image, bbox: Dict[str, float]) -> Dict[str, Any]:
        """Generate segmentation mask for detected item"""
        # TODO: Implement actual segmentation
        x, y, width, height = bbox["x"], bbox["y"], bbox["width"], bbox["height"]
        
        # Placeholder polygon (rectangle)
        polygon = [
            [x, y],
            [x + width, y],
            [x + width, y + height],
            [x, y + height]
        ]
        
        return {
            "polygon": polygon,
            "confidence": 0.9
        }
