services/visual_analysis/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
Pillow>=10.0.0
opencv-python>=4.8.0
albumentations>=1.3.0
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0
pydantic>=2.5.0
typing_extensions>=4.5.0
wandb>=0.16.0
sentry-sdk>=1.32.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
aiofiles==23.2.1
httpx==0.25.2

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
